HOTFIX : menu 수정
- menu 수정

Resolves:
